core/standards/historyinterface/006_14Mar2011.html

<!doctype html>
<html>
	<head>
		<title>Firing popstate after onload, even if there is no pushed/replaced state</title>
		<script type="text/javascript" src="/resources/testharness.js"></script>
		<script type="text/javascript" src="/resources/testharnessreport.js"></script>
		<script type="text/javascript">

//spec (14 March 2011 draft) states that popstate must fire after onload even if there is no pushed/replaced state
var popfired = false;
setup(function () {},{explicit_done:true});
window.addEventListener('popstate',function (e) { popfired = true; },false);
window.onload = function () {
	test(function () {
		assert_false( popfired );
	}, 'popstate event should not fire before onload fires');
	popfired = false;
	setTimeout(function () {
		test(function () {
			assert_true( popfired );
		}, 'popstate event should fire after onload fires, even if there is no pushed/replaced state');
		done();
	},100);
};

		</script>
	</head>
	<body>

		<noscript><p>Enable JavaScript and reload</p></noscript>
		<div id="log"></div>

	</body>
</html>
